\name{chemforms}
\alias{chemforms}
\docType{data}
\title{
Set of exemplary chemical formulas for small molecules.
}
\description{
Vector with character strings of exemplary chemical formulas (pesticides, pharmaceuticals)
}
\usage{data(chemforms)}
\format{
Vector with character strings
}
\examples{
data(chemforms)
chemforms
}
\keyword{datasets}
