\name{isotopes}
\alias{isotopes}
\docType{data}
\title{
Stable isotopes
}
\description{
Dataframe with stable isotopes.
}
\usage{data(isotopes)}
\format{
  A data frame with 302 observations on the following 4 variables.
  \describe{
    \item{\code{element}}{Chemical element} 
    \item{\code{isotope}}{Stable isotopes of an element}
    \item{\code{mass}}{Relative atomic mass}
    \item{\code{abundance}}{Isotopic composition of an element}
  }
}

\source{
\url{http:/physics.nist.gov/cgi-bin/Compositions/stand_alone.pl}
}

\examples{
data(isotopes)
}
\keyword{datasets}
